[REF] cet_website_sale: stock related operations to cet_website_sale_stock
