+++ cet_website_sale/controllers/main.py
@@ -168,27 +168,6 @@ class WebsiteSale(Base):
 
         return response
 
-    @http.route(
-        ["/shop/product/stock_info"], type="json", website=True, auth="public"
-    )
-    def get_product_stock_info(self, **kwargs):
-        """Give a json data structure that contains informations about
-        the available stock for a product variant.
-        """
-        product_id = kwargs.get("id", None)
-        product = request.env["product.product"].sudo().browse(product_id)
-        if product:
-            return {
-                "id": product.id,
-                "virtual_available": product.virtual_available,
-                "inventory_availability": product.inventory_availability,
-                "available_threshold": product.available_threshold,
-                "custom_message": product.custom_message,
-                "cart_qty": product.cart_qty,
-            }
-        else:
-            return {"error": True}
-
     @http.route(["/shop/cart"], type="http", auth="public", website=True)
     def cart(self, access_token=None, revive="", **post):
         response = super().cart(access_token, revive, **post)
@@ -197,7 +176,6 @@ class WebsiteSale(Base):
             sale_order._check_cart_customer_type(
                 request.env.user.partner_id.get_customer_type_id()
             )
-        sale_order._check_cart_available_threshold()
         return response
 
     @http.route(["/shop/checkout"], type="http", auth="public", website=True)
@@ -209,7 +187,7 @@ class WebsiteSale(Base):
             and sale_order._is_restricted(
                 request.env.user.partner_id.get_customer_type_id()
             )
-        ) or sale_order._is_threshold_reached():
+        ):
             return request.redirect("/shop/cart")
         return response
