+++ cet_website_sale/readme/DESCRIPTION.rst
@@ -2,6 +2,9 @@ This modules add lots of customisation for the e-commerce of
 Cycle-en-Terre.
 
 * ...
+* invite login text and custom text
 * display seed growing seasons
-* select a customer type (customer, reseller, food producer, ...)
+* add customer type (customer, reseller, food producer, ...)
+* restrict website features to customer types
+* adds customer types to delivery and picking reports
 * add toggle to include/exclude tax from prices on the ecommerce
