+++ cet_website_sale/static/src/less/design.less
@@ -145,14 +145,6 @@
     }
 }
 
-/* Stock message */
-.oe_stock_messages {
-
-    .stock_message_text {
-        color: @color-orange;
-    }
-}
-
 /* Seedling Calendar */
 .oe_seedling_calendar {
     /* General design for calendar */
