+++ cet_website_sale/static/src/less/product.less
@@ -39,13 +39,6 @@
         font-weight: normal;
     }
 
-    /* Product stock messages */
-    .oe_stock_messages {
-        font-family: "Old Standard TT", Arial, Tahoma, sans-serif;
-        margin-top: 10px;
-        margin-bottom: -15px;
-    }
-
     .oe_quantity_price {
         .oe_p_quantity {
             display: inline-block;
