+++ cet_website_sale/static/src/less/products_item.less
@@ -96,13 +96,6 @@
         padding: 0;
     }
 
-    /* Product stock messages */
-    .oe_stock_messages {
-        font-family: "Old Standard TT", Arial, Tahoma, sans-serif;
-        margin-top: 10px;
-        margin-bottom: -15px;
-    }
-
     /* Product quantity */
     .oe_quantity_price {
         .oe_p_quantity {
