+++ cet_website_sale/views/assets.xml
@@ -22,7 +22,6 @@
         href="/cet_website_sale/static/src/less/product_variants.less"/>
       <link rel="stylesheet" type="text/less"
         href="/cet_website_sale/static/src/less/category_breadcrumb.less"/>
-      <script type="text/javascript" src="/cet_website_sale/static/src/js/stock.js"></script>
       <!-- Taken from oca/e-commerce/website_sale_tax_toggle and adapted for cet_website_sale -->
       <!-- Start of addition -->
       <script type="text/javascript"
