+++ cet_website_sale/views/templates.xml
@@ -63,9 +63,6 @@
         <div t-if='website_sale_order._get_customer_type_warning(clear=False)' class="alert alert-warning">
           <strong>Warning!</strong> <t t-esc='website_sale_order._get_customer_type_warning()'/>
         </div>
-        <div t-if='website_sale_order._get_threshold_warning(clear=False)' class="alert alert-warning">
-          <strong>Warning!</strong> <t t-esc='website_sale_order._get_threshold_warning()'/>
-        </div>
       </t>
     </xpath>
   </template>
@@ -73,11 +70,7 @@
   <template id="cet_website_sale_payment" inherit_id="website_sale.payment">
     <xpath expr="//table[@id='cart_products']" position="after">
       <t t-if='website_sale_order'>
-        <t t-set='warning_threshold' t-value='website_sale_order._get_threshold_warning(clear=False)' />
         <t t-set='warning_customer_type' t-value='website_sale_order._get_customer_type_warning(clear=False)' />
-        <div t-if='warning_threshold' class="alert alert-warning">
-          <strong>Warning!</strong> <t t-esc='website_sale_order._get_threshold_warning()'/>
-        </div>
         <div t-if='warning_customer_type' class="alert alert-warning">
           <strong>Warning!</strong> <t t-esc='website_sale_order._get_customer_type_warning()'/>
         </div>
