+++ cet_website_sale_stock/README.rst
@@ -1,0 +1,63 @@
+=================================
+Cycle en Terre Website Sale Stock
+=================================
+
+.. !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+   !! This file is generated by oca-gen-addon-readme !!
+   !! changes will be overwritten.                   !!
+   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+
+.. |badge1| image:: https://img.shields.io/badge/maturity-Beta-yellow.png
+    :target: https://odoo-community.org/page/development-status
+    :alt: Beta
+.. |badge2| image:: https://img.shields.io/badge/licence-AGPL--3-blue.png
+    :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
+    :alt: License: AGPL-3
+.. |badge3| image:: https://img.shields.io/badge/github-coopiteasy%2Fcycle--en--terre-lightgray.png?logo=github
+    :target: https://github.com/coopiteasy/cycle-en-terre/tree/11.0/cet_website_sale_stock
+    :alt: coopiteasy/cycle-en-terre
+
+|badge1| |badge2| |badge3| 
+
+This modules add stock related customisation for the e-commerce of
+Cycle-en-Terre.
+
+* display warning if product variant stock is below threshold
+* availability threshold is set on variant rather than templateoca
+
+**Table of contents**
+
+.. contents::
+   :local:
+
+Bug Tracker
+===========
+
+Bugs are tracked on `GitHub Issues <https://github.com/coopiteasy/cycle-en-terre/issues>`_.
+In case of trouble, please check there if your issue has already been reported.
+If you spotted it first, help us smashing it by providing a detailed and welcomed
+`feedback <https://github.com/coopiteasy/cycle-en-terre/issues/new?body=module:%20cet_website_sale_stock%0Aversion:%2011.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+
+Do not contact contributors directly about support or help with technical issues.
+
+Credits
+=======
+
+Authors
+~~~~~~~
+
+* Coop IT Easy SCRLfs
+
+Contributors
+~~~~~~~~~~~~
+
+* Rémy Taymans <remy@coopiteasy.be>
+* Robin Keunen <robin@coopiteasy.be>
+
+
+Maintainers
+~~~~~~~~~~~
+
+This module is part of the `coopiteasy/cycle-en-terre <https://github.com/coopiteasy/cycle-en-terre/tree/11.0/cet_website_sale_stock>`_ project on GitHub.
+
+You are welcome to contribute.
