+++ cet_website_sale_stock/__init__.py
@@ -1,0 +1,2 @@
+from . import models
+from . import controllers
