+++ cet_website_sale_stock/controllers/__init__.py
@@ -1,0 +1,1 @@
+from . import main
