+++ cet_website_sale_stock/models/__init__.py
@@ -1,0 +1,1 @@
+from . import sale_order
