+++ cet_website_sale_stock/readme/CONTRIBUTORS.rst
@@ -1,3 +1,3 @@
 * Rémy Taymans <remy@coopiteasy.be>
 * Robin Keunen <robin@coopiteasy.be>
-* Vincent Van Rossem <vincent@coopiteasy.be>
+
