+++ cet_website_sale_stock/readme/DESCRIPTION.rst
@@ -1,0 +1,5 @@
+This modules add stock related customisation for the e-commerce of
+Cycle-en-Terre.
+
+* display warning if product variant stock is below threshold
+* availability threshold is set on variant rather than templateoca
