+++ cet_website_sale_stock/static/src/less/design.less
@@ -1,0 +1,12 @@
+@import url('https://fonts.googleapis.com/css?family=Oswald');
+@import url('https://fonts.googleapis.com/css?family=Old+Standard+TT');
+
+@color-orange: #f0b61c;
+
+/* Stock message */
+.oe_stock_messages {
+
+    .stock_message_text {
+        color: @color-orange;
+    }
+}
