+++ cet_website_sale_stock/static/src/less/product.less
@@ -1,0 +1,9 @@
+/* Main */
+.oe_cet_product_detail .oe_cet_pd_product {
+    /* Product stock messages */
+    .oe_stock_messages {
+        font-family: "Old Standard TT", Arial, Tahoma, sans-serif;
+        margin-top: 10px;
+        margin-bottom: -15px;
+    }
+}
