+++ cet_website_sale_stock/views/templates.xml
@@ -1,0 +1,27 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!-- Copyright 2018 Coop IT Easy SCRLfs
+     License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl). -->
+<odoo>
+
+  <template id="cet_website_sale_cart_lines" inherit_id="website_sale.cart_lines" name="Shopping Cart Lines">
+    <xpath expr="//div[hasclass('js_cart_lines')]" position="after">
+      <t t-if='website_sale_order'>
+        <div t-if='website_sale_order._get_threshold_warning(clear=False)' class="alert alert-warning">
+          <strong>Warning!</strong> <t t-esc='website_sale_order._get_threshold_warning()'/>
+        </div>
+      </t>
+    </xpath>
+  </template>
+
+  <template id="cet_website_sale_payment" inherit_id="website_sale.payment">
+    <xpath expr="//table[@id='cart_products']" position="after">
+      <t t-if='website_sale_order'>
+        <t t-set='warning_threshold' t-value='website_sale_order._get_threshold_warning(clear=False)' />
+        <div t-if='warning_threshold' class="alert alert-warning">
+          <strong>Warning!</strong> <t t-esc='website_sale_order._get_threshold_warning()'/>
+        </div>
+      </t>
+    </xpath>
+  </template>
+
+</odoo>
