+++ cet_website_sale_stock/views/website_sale_templates.xml
@@ -1,0 +1,42 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!-- Copyright 2018-2020 Coop IT Easy SCRLfs
+     License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl). -->
+<odoo>
+
+    <template id="products_item" inherit_id="cet_website_sale.products_item" name="CET Products Item Design - Stock">
+        <xpath expr="//t[@name='cet_product_variant_template']" position="after">
+            <div class="oe_stock_messages hidden">
+                <t t-call="cet_website_sale_stock.stock_messages">
+                    <t t-set="text_class" t-value="'stock_message_text'"/>
+                </t>
+            </div>
+        </xpath>
+    </template>
+
+    <template id="product_detail_section" inherit_id="cet_website_sale.product_detail_section">
+
+        <xpath expr="//p[@name='cet_not_valid_combination_message']" position="after">
+
+            <p t-if="len(product.product_variant_ids) &gt; 1" class="css_not_available_msg bg-danger"
+               style="padding: 15px;">This combination does not exist.
+            </p>
+            <div class="oe_stock_messages hidden">
+                <t t-call="cet_website_sale_stock.stock_messages">
+                    <t t-set="text_class" t-value="'stock_message_text'"/>
+                </t>
+            </div>
+        </xpath>
+    </template>
+
+    <template id="stock_messages">
+        <!--
+          :param: text_class: classes that will be applied to the text.
+        -->
+        <p t-attf-class="{{ text_class }}">
+            <i class="fa fa-exclamation-triangle"></i>
+            Not enough products in
+            stock for now.
+        </p>
+    </template>
+
+</odoo>
